fix: Azure CI build fail on windows
fix the Windows Server 2019 has been retired.
The Windows Server 2019 image has been removed as of 2025-06-30.
fix build break with vs2022 on tgl and ehl platforms.

+++ Platform/TigerlakeBoardPkg/Library/Stage2BoardInitLib/Stage2BoardInitLib.c
@@ -2055,7 +2055,7 @@ UpdateCpuNvs (
   CpuConfigData = (CPU_CONFIG_DATA *)(UINTN)CpuInitDataHob->CpuConfigData;
   CpuSku = GetCpuSku();
 
-  CpuNvs->Cpuid = GetCpuFamily() | GetCpuStepping();
+  CpuNvs->Cpuid = (UINT32)GetCpuFamily() | (UINT32)GetCpuStepping();
   CpuNvs->Revision = CPU_NVS_AREA_REVISION;
   ///
   /// Calculate the number of Oc bins supported. Read in MSR 194h FLEX_RATIO bits (19:17)
